requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.5
ffmpeg-python==0.2.0
future==1.0.0
idna==3.11
pydub==0.25.1
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.3
